cam_note_parser/requirements.txt
asn1crypto==1.3.0
azure-common==1.1.25
azure-core==1.6.0
azure-storage-blob==12.3.2
azure-storage-common==2.1.0
boto3==1.13.26
botocore==1.16.26
certifi==2020.4.5.2
cffi==1.13.2
chardet==3.0.4
cryptography==2.9.2
docutils==0.15.2
DoubleMetaphone==0.1
future==0.18.2
idna==2.9
ijson==2.6.1
isodate==0.6.0
jmespath==0.10.0
msrest==0.6.16
numpy==1.18.5
oauthlib==3.1.0
oscrypto==1.2.0
pandas==1.0.4
probableparsing==0.0.1
probablepeople==0.5.4
pycparser==2.20
pycryptodomex==3.9.7
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-crfsuite==0.9.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
snowflake-connector-python==2.2.7
snowflake-sqlalchemy==1.2.3
SQLAlchemy==1.3.17
urllib3==1.25.9
nltk==3.5